pyproject.toml
[project]
name = "zoekt-mcp"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for Zoekt code search"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Erfan Mirshams", email = "erfan.mirshams@gmail.com" }
]
keywords = ["mcp", "zoekt", "code-search", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastmcp>=3.0.2,<4",
    "pydantic>=2.12.5,<3",
    "requests>=2.32.5,<3",
    "jinja2>=3.1.6,<4",
    "pyyaml>=6.0.3,<7",
    "python-dotenv>=1.2.1,<2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.3",
]

[project.scripts]
zoekt-mcp = "zoekt_mcp.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["zoekt_mcp"]
package-dir = {"zoekt_mcp" = "src"}

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line length is handled by formatter

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
